@@ -9,47 +9,56 @@
   },
   "author": "",
   "devDependencies": {
-    "babel-core": "6.2.4",
-    "babel-loader": "6.2.4",
-    "babel-preset-react": "6.5.0",
-    "babel-preset-es2015": "6.9.0",
-    "css-loader": "0.23.1",
-    "eslint": "2.13.1",
-    "eslint-plugin-react": "5.2.2",
+    "babel-core": "6.26.0",
+    "babel-eslint": "8.0.3",
+    "babel-loader": "7.1.2",
+    "babel-preset-env": "1.6.1",
+    "babel-preset-react": "6.24.1",
+    "critical": "1.1.0",
+    "css-loader": "0.28.7",
+    "eslint": "4.13.1",
+    "eslint-plugin-react": "7.5.1",
     "expect.js": "0.3.1",
-    "file-loader": "0.9.0",
-    "globby": "5.0.0",
+    "file-loader": "0.11.2",
     "gulp": "3.9.1",
-    "gulp-browserify": "0.5.1",
-    "gulp-eslint": "2.0.0",
+    "gulp-clean-css": "3.9.0",
+    "gulp-concat": "2.6.1",
+    "gulp-eslint": "4.0.0",
     "gulp-gzip": "1.4.0",
-    "gulp-clean-css": "2.0.10",
-    "gulp-rename": "1.2.2",
     "gulp-react": "3.1.0",
-    "gulp-shell": "0.5.2",
-    "gulp-sourcemaps": "1.6.0",
-    "gulp-uglify": "1.5.4",
-    "gulp-util": "3.0.7",
-    "jsx-loader": "0.13.2",
-    "less": "2.7.1",
-    "mocha": "2.5.3",
-    "reactify": "1.1.1",
-    "rimraf": "2.5.2",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "webpack": "1.13.1",
-    "webpack-dev-server": "1.14.1"
+    "gulp-rename": "1.2.2",
+    "gulp-sequence": "0.4.6",
+    "gulp-shell": "0.6.3",
+    "gulp-sourcemaps": "2.6.0",
+    "gulp-uglify": "3.0.0",
+    "gulp-util": "3.0.8",
+    "less": "2.7.3",
+    "mocha": "4.0.1",
+    "rimraf": "2.6.1",
+    "style-loader": "0.19.0",
+    "url-loader": "0.5.9",
+    "webpack": "3.10.0",
+    "webpack-dev-server": "2.9.7"
   },
   "dependencies": {
-    "bootstrap": "3.3.6",
-    "handlebars": "4.0.5",
-    "jquery": "3.0.0",
-    "jquery-ui": "1.10.5",
-    "lodash": "4.13.1",
-    "moment": "2.13.0",
-    "react": "15.1.0",
-    "react-dom": "15.1.0",
-    "requirejs": "2.2.0",
-    "select2": "4.0.3"
+    "bootstrap": "3.3.7",
+    "bootstrap-daterangepicker": "2.1.30",
+    "handlebars": "4.0.11",
+    "jquery": "3.2.1",
+    "jquery-ui": "1.12.1",
+    "lodash": "4.17.4",
+    "moment": "2.19.4",
+    "react": "16.2.0",
+    "react-dom": "16.2.0",
+    "datatables.net": "1.10.16",
+    "datatables.net-dt": "1.10.16",
+    "datatables.net-bs": "1.10.16",
+    "datatables.net-buttons": "1.4.2",
+    "datatables.net-buttons-dt": "1.4.2",
+    "datatables.net-buttons-bs": "1.4.2",
+    "datatables-responsive": "1.0.7",
+    "requirejs": "2.3.5",
+    "select2": "4.0.4",
+    "select2-bootstrap-theme": "0.1.0-beta.10"
   }
 }
